Rename remaining occurences
