+++ src/py/flwr/simulation/ray_transport/ray_client_proxy_test.py
@@ -22,7 +22,7 @@ from typing import Dict, List, Tuple, Type
 import ray
 
 from flwr.client import Client, NumPyClient
-from flwr.client.clientapp import ClientApp
+from flwr.client.client_app import ClientApp
 from flwr.common import (
     Config,
     ConfigsRecord,
